fix: move the log package out of cmd package that can be used by other packages outside cmd
