+++ pkg/gits/bitbucket_server.go
@@ -12,7 +12,7 @@ import (
 
 	bitbucket "github.com/gfleury/go-bitbucket-v1"
 	"github.com/jenkins-x/jx/pkg/auth"
-	"github.com/jenkins-x/jx/pkg/jx/cmd/log"
+	"github.com/jenkins-x/jx/pkg/log"
 	"github.com/jenkins-x/jx/pkg/util"
 )
