+++ pkg/jx/cmd/create_cluster_minikube.go
@@ -11,9 +11,10 @@ import (
 	"time"
 
 	"fmt"
-	"github.com/jenkins-x/jx/pkg/jx/cmd/log"
+
 	"github.com/jenkins-x/jx/pkg/jx/cmd/templates"
 	cmdutil "github.com/jenkins-x/jx/pkg/jx/cmd/util"
+	"github.com/jenkins-x/jx/pkg/log"
 	"github.com/jenkins-x/jx/pkg/util"
 	"github.com/spf13/cobra"
 	"gopkg.in/AlecAivazis/survey.v1"
