+++ pkg/jx/cmd/step_next_version.go
@@ -16,9 +16,9 @@ import (
 	"github.com/blang/semver"
 	version "github.com/hashicorp/go-version"
 	"github.com/jenkins-x/jx/pkg/gits"
-	"github.com/jenkins-x/jx/pkg/jx/cmd/log"
 	"github.com/jenkins-x/jx/pkg/jx/cmd/templates"
 	cmdutil "github.com/jenkins-x/jx/pkg/jx/cmd/util"
+	"github.com/jenkins-x/jx/pkg/log"
 	"github.com/spf13/cobra"
 )
