+++ pkg/jx/cmd/update_cluster_gke_terraform.go
@@ -7,14 +7,15 @@ import (
 
 	os_user "os/user"
 
-	"github.com/jenkins-x/jx/pkg/jx/cmd/gke"
-	"github.com/jenkins-x/jx/pkg/jx/cmd/log"
-	"github.com/jenkins-x/jx/pkg/jx/cmd/templates"
-	cmdutil "github.com/jenkins-x/jx/pkg/jx/cmd/util"
-	"github.com/spf13/cobra"
-	"gopkg.in/AlecAivazis/survey.v1"
 	"os"
 	"path/filepath"
+
+	"github.com/jenkins-x/jx/pkg/jx/cmd/gke"
+	"github.com/jenkins-x/jx/pkg/jx/cmd/templates"
+	cmdutil "github.com/jenkins-x/jx/pkg/jx/cmd/util"
+	"github.com/jenkins-x/jx/pkg/log"
+	"github.com/spf13/cobra"
+	"gopkg.in/AlecAivazis/survey.v1"
 )
 
 // CreateClusterOptions the flags for running create cluster
