+++ pkg/kube/cluster_status.go
@@ -2,7 +2,8 @@ package kube
 
 import (
 	"fmt"
-	"github.com/jenkins-x/jx/pkg/jx/cmd/log"
+
+	"github.com/jenkins-x/jx/pkg/log"
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
